ci: move prek to fingerprinted uv-cache artifacts
- Replace prebuilt-image bootstrap with uv-cache release-asset workflow and fingerprint gating.
- Use compute/build helpers (`compute_uv_fingerprint.py`, `build_and_push_uv_cache.sh`) and document cache refresh in CONTRIBUTING.
- Harden containerized Prek workflow for deterministic execution (git installed before checkout, explicit safe.directory, POSIX-safe shell checks).
- Enforce fingerprint-only cache restore by removing moving `*-current.tar.zst` fallback from restore and publish paths.
- Keep Prek dependency scope at `--all-extras --group dev` to preserve megatron coverage in CI.
- Store full uv cache as fingerprinted chunked release assets (`.tar.zst.part-###`) so CI reuses wheel/build payloads without GH single-asset size limits.
- Download cache parts with parallelism 8 in CI restore path, then reassemble in deterministic order.
- Bump cache fingerprint schema/layout contract for the chunked-asset format.
- Keep immutable cache retention policy (latest 4 fingerprints).
