+++ pyproject.toml
@@ -40,6 +40,17 @@ backend = [
     "gql<4",
     "vllm==0.15.1 ; sys_platform == 'linux'",
 ]
+megatron = [
+    "torch>=2.8.0",
+    "apex",
+    "transformer-engine==2.11.0",
+    "transformer-engine-cu12==2.11.0",
+    "transformer-engine-torch==2.11.0",
+    "megatron-core==0.15.2",
+    "megatron-bridge==0.2.0rc6",
+    "nvidia-ml-py==13.580.82",
+    "ml-dtypes>=0.5.0 ; python_full_version < '3.13'",
+]
 
 langgraph = [
     "langchain-core>=0.3.51",
@@ -111,7 +122,16 @@ required-version = ">=0.6.15"
 # Override numpy to <2.0 for compatibility with megatron-core in the training
 # environment. vLLM 0.15.1 pulls opencv-python-headless>=4.13 which wants
 # numpy>=2 on Python 3.9+, but megatron-core requires numpy<2.
-override-dependencies = ["numpy<2"]
+override-dependencies = ["transformer-engine>=2.11.0", "numpy<2"]
+exclude-dependencies = ["pynvml"]
+no-build-isolation-package = ["apex", "transformer-engine-torch", "nv-grouped-gemm"]
+
+[tool.uv.extra-build-dependencies]
+apex = ["torch>=2.8.0"]
+transformer-engine-torch = ["torch>=2.8.0"]
+
+[tool.uv.extra-build-variables]
+apex = { APEX_CPP_EXT = "1", APEX_CUDA_EXT = "1", APEX_FAST_LAYER_NORM = "1", APEX_PARALLEL_BUILD = "1", NVCC_APPEND_FLAGS = "--threads 1" }
 
 [tool.ty.environment]
 python-version = "3.11"
@@ -173,7 +193,7 @@ dev = [
     "pytest>=8.4.1",
     "nbval>=0.11.0",
     "pytest-xdist>=3.8.0",
-    "ty>=0.0.14",
+    "ty==0.0.14",
     "pytest-asyncio>=1.1.0",
     "duckdb>=1.0.0",
     "pyarrow>=15.0.0",
@@ -183,3 +203,4 @@ dev = [
 
 [tool.uv.sources]
 panza = { git = "https://github.com/corbt/panza.git" }
+apex = { git = "https://github.com/NVIDIA/apex.git", branch = "25.09" }
